LayoutTests/imported/w3c/web-platform-tests/permissions-policy/payment-allowed-by-permissions-policy-attribute-redirect-on-load.https.sub-expected.txt

CONSOLE MESSAGE: Permission policy 'Payment' check failed for document with origin 'https://www1.web-platform.test:9443'.

PASS permissions policy allow="payment" allows same-origin navigation in an iframe.
PASS permissions policy allow="payment" disallows cross-origin navigation in an iframe.
